+++ examples/basic-rml/src/app.ts
@@ -1,6 +1,0 @@
-import { runApp, IAppConfig } from 'ice';
-
-const appConfig: IAppConfig = {
-};
-
-runApp(appConfig);