+++ examples/basic-ssr-with-lazy-load/src/app.tsx
@@ -1,11 +1,23 @@
-import { runApp, IAppConfig, config, request } from 'ice';
+import { runApp, IAppConfig, config } from 'ice';
 import React from 'react';
+
+const delay = (time) => new Promise<void>((resolve) => setTimeout(() => resolve(), time));
 
 const appConfig: IAppConfig = {
   app: {
     getInitialData: async () => {
-      const res = await request('/user');
-      return res;
+      // const res = await request('/user');
+      // return res;
+      await delay(1500);
+
+      return {
+        data: {
+          user: {
+            name: 'Jack Ma',
+            id: 10001,
+          }
+        },
+      };
     }
   },
   router: {