+++ examples/basic-ssr-with-lazy-load/src/pages/About/index.tsx
@@ -1,16 +1,16 @@
 import React from 'react';
-import { Link, logger, Helmet } from 'ice';
+import { Link, logger, Head } from 'ice';
 
 const About = (props) => {
   logger.info('About props', props);
   return (
     <>
-      <Helmet>
+      <Head>
         <meta charSet="utf-8" />
         <title>About</title>
         <meta name="keywords" content="About Keywords" />
         <meta name="description" content="About Description" />
-      </Helmet>
+      </Head>
       <h2>{props.title}</h2>
       <Link to="/">home</Link>
     </>