+++ packages/plugin-helmet/src/index.ts
@@ -2,8 +2,14 @@ export default ({ applyMethod }) => {
   applyMethod('addExport', {
     source: 'react-helmet',
     specifier: '{ Helmet }',
-    exportName: 'Helmet',
+    exportName: `
+    Helmet as Head,
+    // @deprecated
+    Helmet`,
     importSource: 'react-helmet',
     exportMembers: ['Helmet'],
+    alias: {
+      'Head': 'Helmet'
+    }
   });
 };