+++ packages/plugin-icestark/src/babelPluginMicroapp.ts
@@ -1,7 +1,9 @@
 import * as t from '@babel/types';
 
 const templateIfStatement = 'if (!isInIcestark()) {}';
+
 const templateExportStatement = `
+setLibraryName(LIBRARY);
 export const mount = (props) => {
   APP_CALLEE(APP_CONFIG);
 };
@@ -11,6 +13,9 @@ export const unmount = ({ container, customProps }) => {
   } else {
     ReactDOM.unmountComponentAtNode(container);
   }
+};
+export const bootstrap = () => {
+  console.log('bootstrap');
 };`;
 const templateModeStatement = `
 if (typeof window !== 'undefined' && window.ICESTARK && window.ICESTARK.loadMode && window.ICESTARK.loadMode !== 'umd') {
@@ -23,7 +28,7 @@ const getUid = () => {
   return () => `_APP_CONFIG${(uid++) || ''}`;
 };
 
-export default (api, { entryList }) => {
+export default (api, { entryList, libraryName }) => {
   const namespaceSpecifier: string[] = [];
   const importSpecifier: string[] = [];
   let configIdentifier: string;
@@ -60,13 +65,20 @@ export default (api, { entryList }) => {
               } else if (t.isIdentifier(item.source, { value: '@ice/stark-app'})) {
                 starkappStatement = true;
                 let importIsInIcestark = false;
+                let importSetLibraryName = false;
                 item.specifiers.forEach((value) => {
                   if (t.isImportSpecifier(value) && t.isIdentifier(value.local, { name: 'isInIcestark'})) {
                     importIsInIcestark = true;
                   }
+                  if (t.isImportSpecifier(value) && t.isIdentifier(value.local, { name: 'setLibraryName'})) {
+                    importSetLibraryName = true;
+                  }
                 });
                 if (!importIsInIcestark) {
                   item.specifiers.push(t.importSpecifier(t.identifier('isInIcestark'), t.identifier('isInIcestark')));
+                }
+                if (!importSetLibraryName) {
+                  item.specifiers.push(t.importSpecifier(t.identifier('setLibraryName'), t.identifier('setLibraryName')));
                 }
               // check import ReactDOM from 'react-dom';
               } else if (t.isIdentifier(item.source, { value: 'react-dom'})) {
@@ -87,7 +99,10 @@ export default (api, { entryList }) => {
           // import @ice/stark-app
           if (!starkappStatement) {
             const starkappImport = t.importDeclaration(
-              [t.importSpecifier(t.identifier('isInIcestark'), t.identifier('isInIcestark'))],
+              [
+                t.importSpecifier(t.identifier('isInIcestark'), t.identifier('isInIcestark')),
+                t.importSpecifier(t.identifier('setLibraryName'), t.identifier('setLibraryName'))
+              ],
               t.stringLiteral('@ice/stark-app'),
             );
             lastImportIndex += 1;
@@ -158,6 +173,7 @@ export default (api, { entryList }) => {
             const astExport = api.template(templateExportStatement)({
               APP_CONFIG: configIdentifier,
               APP_CALLEE: callIdentifier || identifierCallee,
+              LIBRARY: t.stringLiteral(libraryName),
             });
             nodePath.insertAfter(astExport);
             replaced = true;